Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -2,10 +2,10 @@
 
 int main(){
 
-float A;
+double A;
 scanf("%lf", &A);
 ;
-printf("You entered: %lf",A);
+printf("You entered: %lf", A);
 
     return 0;
 }
